Extract InputSystem from the desktop domain. It doesn't deppend directly on Desktop Window anymore
