+++ Spatial.Core/include/spatial/input/InputSystem.h
@@ -1,0 +1,36 @@
+#pragma once
+
+#include <spatial/desktop/PlatformEvent.h>
+#include <spatial/input/InputState.h>
+#include <spatial/common/Signal.h>
+
+namespace spatial::desktop
+{
+
+class InputSystem
+{
+  public:
+	InputSystem();
+
+	InputSystem(const InputSystem& other) = delete;
+	InputSystem& operator=(const InputSystem& w) = delete;
+
+	void onUpdateFrame(float delta);
+
+	void onEndFrame();
+
+	void onEvent(const KeyEvent& event);
+
+	void onEvent(const MouseMovedEvent& event);
+
+	auto& getOnChangeStateSignal()
+	{
+		return mOnChangeStateSignal;
+	}
+
+  private:
+	InputState mInputState;
+	Signal<const InputState&> mOnChangeStateSignal;
+};
+
+} // namespace spatial::desktop
